requirements.txt
torch
torchvision
gradio
numpy
pandas
matplotlib
tqdm
scikit-learn
pillow